% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{antidep}
\alias{antidep}
\title{wide format of continuous response of antidepressant data.}
\format{
A data frame with 172 rows and 6 variables:
\describe{
\item{PID}{Patient ID}
\item{tx}{Treatment, 1 for treated and 0 for placebo}
\item{y0}{HADM-17 measurement at the baseline}
\item{y1, y2, y4, y6}{Change score of HADM-17 measurement at the post-baseline week 1, 2, 4, and 6.}
}
}
\source{
https://www.lshtm.ac.uk/research/centres-projects-groups/missing-data#dia-missing-data
}
\usage{
antidep
}
\description{
A data set containing the treatment and continuous responses measured at baseline and 4 post-baseline visits
}
\keyword{datasets}
