% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_class.R
\name{get_class}
\alias{get_class}
\title{obtain ordinal results based on eta and gamma from MCMC}
\usage{
get_class(MCMC, impvar, eta_name, rev = FALSE, seed = NULL)
}
\arguments{
\item{MCMC}{an matrix of MCMC samples.}

\item{impvar}{a name of imputation variable.}

\item{eta_name}{a name of eta in mcmc samples}

\item{rev}{logical. Reverse order or not.}

\item{seed}{optional. A seed value for randomness.}
}
\description{
obtain ordinal results based on eta and gamma from MCMC
}
