% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Mlist.R
\name{get_Mlist}
\alias{get_Mlist}
\title{Prepare imputation-model-related information}
\usage{
get_Mlist(object)
}
\arguments{
\item{object}{object inheriting from class \code{remoid}.}
}
\value{
a list include raw data, imputation models, model types, fixed effects,
random effects if any, reference categories corresponding to categorical
variables in models, and interaction terms.
}
\description{
Internal function to extract information of imputation models.
}
