% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modeltypes_custom.R
\name{get_models_custom}
\alias{get_models_custom}
\title{Specify the default (imputation) model types}
\usage{
get_models_custom(fixed, random = NULL, data, auxvars = NULL,
  timevar = NULL, no_model = NULL, models = NULL, model_order = NULL,
  analysis_type = NULL, warn = TRUE)
}
\arguments{
\item{fixed}{a two sided formula describing the fixed-effects part of the
model (see \code{\link[stats]{formula}})}

\item{random}{only for multi-level models:
a one-sided formula of the form \code{~x1 + ... + xn | g},
where \code{x1 + ... + xn} specifies the model for the random
effects and \code{g} the grouping variable}

\item{data}{a \code{data.frame} containing the original data
(more details below)}

\item{auxvars}{optional; one-sided formula of variables that should be used
as predictors in the imputation procedure (and will be imputed
if necessary) but are not part of the analysis model(s).}

\item{timevar}{name of the variable indicating the time of the measurement
of a time-varying covariate in a proportional hazards survival mode.}

\item{no_model}{optional; vector of names of variables for which no model
should be specified. Note that this is only possible for
completely observed variables and implies the assumptions
of independence between the excluded variable and the
incomplete variables.}

\item{models}{optional; named vector specifying the types of models for
(incomplete) covariates.
This arguments replaces the argument \code{meth} used in
earlier versions.
If \code{NULL} (default) models will be determined
automatically based on the class of the respective columns of
\code{data}.}

\item{model_order}{optional; manually specify an order for imputation models.}

\item{analysis_type}{type of analysis, e.g. \code{clm} for ordinal variable.}

\item{warn}{logical, should warnings be displayed?}
}
\description{
Modified from codes in JointAI to allow users customizing the orders of sequential models
Original code set the order of models based on number of missing values from each variable
Now, the order of models can be specified through \code{model_order}.
}
\note{
the following scenario gives the sequential imputation models in an order of
y0, y1, y3, y2, y4,..., which is based on 'nmis' variable. The expected order
would be y0, y1, y2, y3, y4,.... To specify the order, what we need is just set
model_order = c('y0', 'y1', 'y2', 'y3', 'y4', 'y5')

rowID   out lvl nmis nlev ordered type L1
8 rowID1 FALSE   1  428    3    TRUE  clm y5
7 rowID1 FALSE   1  426    4    TRUE  clm y4
5 rowID1 FALSE   1  423    4    TRUE  clm y2
6 rowID1 FALSE   1   63    4    TRUE  clm y3
4 rowID1 FALSE   1   11    4    TRUE  clm y1
3 rowID1 FALSE   1    3    4    TRUE  clm y0
}
