% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.models.R
\name{list.models}
\alias{list.models}
\title{Listing the sequence of models used for imputation}
\usage{
list.models(object, details = FALSE, print = TRUE)
}
\arguments{
\item{object}{an object of class \code{remiod}}

\item{details}{logical. Default is FALSE, where listing all models in formula format. If TRUE,
details of each models will be presented.}

\item{print}{logical. Default is TRUE to print all imputation models or detailed imputation models.}
}
\description{
Listing the sequence of models used for imputation
}
