% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.remiod.R
\name{print.summary.remiod}
\alias{print.summary.remiod}
\title{print summary outputs}
\usage{
\method{print}{summary.remiod}(x, digits = max(3, .Options$digits - 4), ...)
}
\arguments{
\item{x}{an object of class \code{summary.remiod}}
}
\description{
print summary outputs
}
