% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schizow}
\alias{schizow}
\title{wide format of ordinal response of Schizophrenia data.}
\format{
A data frame with 437 rows and 5 variables:
\describe{
\item{tx}{treatment, 1 for treated and 0 for placebo}
\item{y0}{ordinal response at the baseline}
\item{y1, y3, y6}{ordinal response at the post-baseline week 1, 3, and 6.}
}
}
\source{
long-to-wise tranformation of schizo data, i.e. schizow = reshape2::dcast(schizo, id + tx ~ week, value.var = "imps79o")
}
\usage{
schizow
}
\description{
A dataset containing the treatment and ordinal responses measured at baseline and 3 post-baseline visits
}
\keyword{datasets}
