% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.remiod.R
\name{summary.remiod}
\alias{summary.remiod}
\title{Summarize the results from an object of class JointAI}
\usage{
\method{summary}{remiod}(Object, start = NULL, end = NULL, thin = NULL,
  quantiles = c(0.025, 0.975), exclude_chains = NULL, outcome = NULL,
  warn = TRUE, mess = TRUE, ...)
}
\arguments{
\item{start}{the first iteration of interest
(see \code{\link[coda]{window.mcmc}})}

\item{end}{the last iteration of interest
(see \code{\link[coda]{window.mcmc}})}

\item{quantiles}{posterior quantiles}

\item{exclude_chains}{optional vector of the index numbers of chains that
should be excluded}

\item{digits}{the minimum number of significant digits to be printed in
values.}
}
\description{
Obtain and print the \code{summary}, (fixed effects) coefficients
(\code{coef}) and credible interval (\code{confint}).
}
