% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tang_MI_RB.R
\name{tang_MI_RB}
\alias{tang_MI_RB}
\title{Implement the algorithm proposed by Tang}
\usage{
tang_MI_RB(object, dtimp, treatment, method = "MAR", delta = 0,
  ord_cov_dummy = FALSE, exclude_chains = NULL, include = FALSE)
}
\arguments{
\item{object}{object inheriting from class 'remoid'}

\item{dtimp}{imputed complete data sets from \code{remiod} function.}

\item{treatment}{treatment variable.}

\item{method}{a method for obtaining multiple-imputed dataset. Options include
\code{MAR}, \code{J2R}, \code{CR}, and \code{delta} adjustment.
Default is MAR.}

\item{delta}{specific value used for Delta adjustment, applicable only
for method="delta".}

\item{ord_cov_dummy}{optional. specify whether ordinal variables should be treated as
categorical variables or continuous variables when they are
included as covariates in the sequential imputation models.
Default is \code{TRUE}, dummy variables will be created accordingly.}

\item{exclude_chains}{optional vector of the index numbers of chains that
should be excluded}

\item{include}{logical, if TRUE, raw data will be included in imputed data sets
with imputation ID = 0.}
}
\description{
Implement the algorithm proposed by Tang
}
\keyword{internal}
