% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMI.R
\name{updateMI}
\alias{updateMI}
\title{Apply a MI method following initial run of \code{remoid} function}
\usage{
updateMI(object, method = c("MAR", "J2R", "CR", "delta"), delta = 0,
  mi.setting = NULL, mess = TRUE)
}
\arguments{
\item{object}{object inheriting from class \code{remoid}.}

\item{method}{a method for obtaining multiple-imputed dataset. Options include
\code{MAR}, \code{J2R}, \code{CR}, and \code{delta} adjustment.
Default is MAR.}

\item{delta}{specific value used for Delta adjustment, applicable only
for method="delta".}

\item{mi.setting}{a list of arguments for extracting MI data set, which
will be used to update the one in \code{remoid} object.
Default is NULL, meaning no update to the mi.setting
in \code{remoid} object.}

\item{mess}{logical; should messages be given? Default is TRUE.}
}
\description{
Apply a MI method following initial run of \code{remoid} function
}
\keyword{internal}
